Move extra keys informations in a new class
